Started restructure, need to connect everything
