+++ client/src/components/LayoutMain/LayoutMain.js
@@ -1,0 +1,20 @@
+import React from 'react'
+
+const LayoutMain = props => {
+  return (
+    <div className="layout-main">
+      <div className="layout-main__navbar">
+      
+      </div>
+      <div className="layout-main__body">
+      
+      </div>
+      <div className="layout-main__sidebar">
+      
+      </div>
+      <div className="layout-main__footer">
+      
+      </div>
+    </div>
+  )
+}
