--- client/src/pages/Builder/BuilderPage.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-
-class BuilderPage extends Component {
-  state = {};
-
-  render() {
-    return (
-      <div className="builder__page">
-        Builder Page
-      </div>
-    )
-  }
-}
-
-export default BuilderPage;
