--- client/src/pages/Calendar/CalendarView.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-
-class CalendarView extends Component {
-  state = {};
-
-  render() {
-    return (
-      <div className="calendar__view">
-        Calendar View Page
-      </div>
-    )
-  }
-}
-
-export default CalendarView;
