--- client/src/pages/Library/LibraryAddAlbums.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-class LibraryAddAlbums extends Component {
-  render() {
-    return (
-      <div>
-        Library Add Albums
-      </div>
-    )
-  }
-}
-
-export default LibraryAddAlbums;
