--- client/src/pages/Library/LibraryAddArtists.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-class LibraryAddArtists extends Component {
-  render() {
-    return (
-      <div>
-        Library Add Artists
-      </div>
-    )
-  }
-}
-
-export default LibraryAddArtists;
