--- client/src/pages/Library/LibraryAddTracks.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-class LibraryAddTracks extends Component {
-  render() {
-    return (
-      <div>
-        Library Add Tracks        
-      </div>
-    )
-  }
-}
-
-export default LibraryAddTracks;
