--- client/src/pages/Report/ReportPage.js
@@ -1,26 +1,0 @@
-import React, { Component } from 'react'
-
-class ReportPage extends Component {
-  state = {};
-
-  render() {
-    return (
-      <div className="report__page">
-        Report Page
-
-       <div className="footer__demo">
-
-       <footer className="footer">
-
-       <p className="footer-text">Powered By &nbsp;</p>
-       <img id="footer-conrad-logo" src="https://i.postimg.cc/dVFWp5Fq/Comrad-Logo-13.png"/>
-       
-       </footer>
-
-       </div>   
-      </div>
-    )
-  }
-}
-
-export default ReportPage;
