--- client/src/pages/User/UserEdit.js
@@ -1,11 +1,0 @@
-import React, { Component } from 'react'
-
-export default class UserEdit extends Component {
-  render() {
-    return (
-      <div>
-        User Edit Page
-      </div>
-    )
-  }
-}
