+++ client/src/pages/UserAddPage/index.js
@@ -1,0 +1,1 @@
+export { default } from './UserAddPage'
