+++ client/src/pages/UserSearchPage/index.js
@@ -1,0 +1,1 @@
+export { default } from './UserSearchPage'
