--- client/src/routes/Admin.js
@@ -1,22 +1,0 @@
-import React, { Component } from 'react'
-import { Route } from 'react-router-dom'
-
-import AdminPage from '../pages/Admin/AdminPage';
-
-class Admin extends Component {
-  state = {}
-
-  render() {
-    const { url } = this.props.match;
-
-    return (
-      <main className="admin">
-        <section className="admin__body">
-          <Route exact path={`${url}/`} component={AdminPage} />
-        </section>
-      </main>
-    )
-  }
-}
-
-export default Admin
