+++ client/src/routes/LibraryRoutes/index.js
@@ -1,0 +1,1 @@
+export { default } from './LibraryRoutes'
