+++ client/src/routes/UserRoutes/index.js
@@ -1,0 +1,1 @@
+export { default } from './UserRoutes'
