+++ client/src/utils/serviceWorker/registerServiceWorker.js
