+++ package.json
@@ -14,7 +14,7 @@
   "scripts": {
     "start": "node index.js",
     "server": "nodemon server/index.js --ignore tests --delay 1",
-    "client": "npm run start --prefix client",
+    "client": "npm run start --parefix client",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
     "build": "NPM_CONFIG_PRODUCTION=false npm install --prefix client && npm run build --prefix client",
     "heroku-postbuild": "npm run build",
